Don't delete the canary

+++ app/commands/tooling_job/delete_from_efs.rb
@@ -4,6 +4,7 @@ class ToolingJob::DeleteFromEFS
   initialize_with :efs_dir
 
   def call
+    return if efs_dir == "/mnt/efs/tooling_jobs/canary"
     return unless Dir.exist?(efs_dir)
 
     FileUtils.rm_rf(efs_dir)
